Activity Feed Notification created

+++ lib/widgets/post.dart
@@ -168,32 +168,41 @@ class _PostState extends State<Post> {
   }
 
   addLikeToActivityFeed() {
-    activityFeedRef
-        .document(ownerId)
-        .collection('feedItems')
-        .document(postId)
-        .setData({
-      'type': 'like',
-      'username': currentUser.username,
-      'userId': currentUser.id,
-      'userProfileImg': currentUser.photoUrl,
-      'postId': postId,
-      'mediaUrl': mediaUrl,
-      'timestamp': timestamp,
-    });
+    /* add a notification to the postOwner's activity feed only if 
+      comment made by OTHER user (to avoid getting notification for our
+      own like)*/
+    bool isNotPostOwner = currentUserId != ownerId;
+    if (isNotPostOwner) {
+      activityFeedRef
+          .document(ownerId)
+          .collection('feedItems')
+          .document(postId)
+          .setData({
+        'type': 'like',
+        'username': currentUser.username,
+        'userId': currentUser.id,
+        'userProfileImg': currentUser.photoUrl,
+        'postId': postId,
+        'mediaUrl': mediaUrl,
+        'timestamp': timestamp,
+      });
+    }
   }
 
   removeLikeFromActivityFeed() {
-    activityFeedRef
-        .document(ownerId)
-        .collection('feedItems')
-        .document(postId)
-        .get()
-        .then((doc) {
-      if (doc.exists) {
-        doc.reference.delete();
-      }
-    });
+    bool isNotPostOwner = currentUserId != ownerId;
+    if (isNotPostOwner) {
+      activityFeedRef
+          .document(ownerId)
+          .collection('feedItems')
+          .document(postId)
+          .get()
+          .then((doc) {
+        if (doc.exists) {
+          doc.reference.delete();
+        }
+      });
+    }
   }
 
   buildPostImage() {
